% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_export_tasks}
\alias{ec2_describe_export_tasks}
\title{Describes the specified export instance tasks or all of your export
instance tasks}
\usage{
ec2_describe_export_tasks(ExportTaskIds = NULL, Filters = NULL)
}
\arguments{
\item{ExportTaskIds}{The export task IDs.}

\item{Filters}{the filters for the export tasks.}
}
\description{
Describes the specified export instance tasks or all of your export instance tasks.

See \url{https://paws-r.github.io/docs/ec2/describe_export_tasks.html} for full documentation.
}
\keyword{internal}
