% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_service_deployments}
\alias{lightsail_get_container_service_deployments}
\title{Returns the deployments for your Amazon Lightsail container service}
\usage{
lightsail_get_container_service_deployments(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the container service for which to return deployments.}
}
\description{
Returns the deployments for your Amazon Lightsail container service

See \url{https://paws-r.github.io/docs/lightsail/get_container_service_deployments.html} for full documentation.
}
\keyword{internal}
