% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_peer_vpc}
\alias{lightsail_peer_vpc}
\title{Peers the Lightsail VPC with the user's default VPC}
\usage{
lightsail_peer_vpc()
}
\description{
Peers the Lightsail VPC with the user's default VPC.

See \url{https://paws-r.github.io/docs/lightsail/peer_vpc.html} for full documentation.
}
\keyword{internal}
