% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_start_instance}
\alias{lightsail_start_instance}
\title{Starts a specific Amazon Lightsail instance from a stopped state}
\usage{
lightsail_start_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance (a virtual private server) to start.}
}
\description{
Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the \verb{reboot instance} operation.

See \url{https://paws-r.github.io/docs/lightsail/start_instance.html} for full documentation.
}
\keyword{internal}
